dbを検証中、第一弾修了
