+++ old/src/db/basic40_special_cards.xlsx
