+++ src/dmps_cards/basic/basic_red_20200513.csv
@@ -1,0 +1,13 @@
+id,name,tribe1,tribe2,rarity,cost ,culture1,culture2,category,series
+4EAD,クック・ポロン, ファイアー・バード, , BASIC,2, 火, , クリーチャー, BASIC
+ZAAD,喧嘩屋タイラー, ヒューマノイド, , BASIC,2, 火, , クリーチャー, BASIC
+AAAE,双神兵カミカゼ, アーマロイド, , BASIC,2, 火, , クリーチャー, BASIC
+V4AD,不死身男爵ボーグ, ヒューマノイド, , BASIC,2, 火, , クリーチャー, BASIC
+GIAE,炎獣兵マグナム・ブルース, ドラゴノイド, , BASIC,3, 火, , クリーチャー, BASIC
+SYAE,幻竜砲, 呪文, , BASIC,3, 火, , 呪文, BASIC
+DEAE,爆炎野郎ジョー, ヒューマノイド, , BASIC,3, 火, , クリーチャー, BASIC
+ZAAE,機神装甲ヴァルディオス, ヒューマノイド, , BASIC,4, 火, , クリーチャー, BASIC
+MQAE,撃墜団長メッツァー, ヒューマノイド, , BASIC,4, 火, , クリーチャー, BASIC
+JMAE,掃討兵バーニング・ヘル, ドラゴノイド, , BASIC,4, 火, , クリーチャー, BASIC
+V4AE,トルネード・フレーム, 呪文, , BASIC,5, 火, , 呪文, BASIC
+PUAE,ボルシャック・ドラゴン, アーマード・ドラゴン, , BASIC,6, 火, , クリーチャー, BASIC
