+++ src/dmps_cards/prize/prize_20200513.csv
@@ -1,0 +1,10 @@
+id,name,tribe1,tribe2,rarity,cost ,culture1,culture2,category,series
+SYBL,ブラッディ・イヤリング, ブレインジャッカー, , BASIC,2, 闇, , クリーチャー, BASIC
+PUA4,デーモン・ハンド, 呪文, , VR,6, 闇, , 呪文,超獣の始動
+MQA5,クック・ポロン, ファイアー・バード, , BASIC,2, 火, , クリーチャー,超獣の始動
+GIA4,バースト・ショット, 呪文, , VR,6, 火, , 呪文,超獣の始動
+DEA4,ホーリー・スパーク, 呪文, , VR,6, 光, , 呪文,超獣の始動
+V4BL,フェアリー・ライフ, 呪文, , BASIC,2, 自然, , 呪文, BASIC
+JMA4,ナチュラル・トラップ, 呪文, , VR,6, 自然, , 呪文,超獣の始動
+MQA4,アクア・サーファー, リキッド・ピープル, , VR,6, 水, , クリーチャー,超獣の始動
+PUBL,アクア・ハルカス, リキッド・ピープル, , BASIC,3, 水, , クリーチャー, BASIC
