+++ src/dmps_cards/secret/secret_20200513.csv
@@ -1,0 +1,8 @@
+id,name,tribe1,tribe2,rarity,cost ,culture1,culture2,category,series
+DYAP,悪魔神バロム, デーモン・コマンド, , SR,8, 闇, , クリーチャー,超獣の始動
+WQAK,クリスタル・パラディン(SEC), リキッド・ピープル, , SR,4, 水, , クリーチャー,超獣の始動
+ZUBJ,幻想妖精カチュア, スノーフェアリー, , SR,5, 自然, , クリーチャー,伝説の再誕
+NEAG,精霊王アルカディアス, エンジェル・コマンド, , SR,6, 光, , クリーチャー,超獣の始動
+AUAY,大勇者「ふたつ牙」, ビーストフォーク, , SR,6, 自然, , クリーチャー,超獣の始動
+KAAT,ボルメテウス・ホワイト・ドラゴン, アーマード・ドラゴン, , SR,7, 火, , クリーチャー,超獣の始動
+G4BI,超龍バジュラ, アーマード・ドラゴン, , SR,7, 火, , クリーチャー,伝説の再誕
