+++ src/dmps_cards/series1/series1_black_20200513.csv
@@ -1,0 +1,45 @@
+id,name,tribe1,tribe2,rarity,cost,culture1,culture2,category,series
+4EAR,暴虐虫タイラント・ワーム, パラサイトワーム, , C,1, 闇, , クリーチャー,超獣の始動
+AAAS,ねじれる者ボーン・スライム, リビング・デッド, , C,1, 闇, , クリーチャー,超獣の始動
+DEAT,ブラック・スレイヤー, 呪文, , C,1, 闇, , 呪文,超獣の始動
+DEAS,うごめく者ボーン・グール, リビング・デッド, , C,2, 闇, , クリーチャー,超獣の始動
+GIAS,さまよう者ブレイン・イーター, リビング・デッド, , C,2, 闇, , クリーチャー,超獣の始動
+4EAQ,飛行男, ヘドリアン, , UC,2, 闇, , クリーチャー,超獣の始動
+GIAT,ゴースト・タッチ, 呪文, , C,2, 闇, , 呪文,超獣の始動
+V4AR,ダーク・リバース, 呪文, , UC,2, 闇, , 呪文,超獣の始動
+JMAS,ギガゾウル, キマイラ, , C,3, 闇, , クリーチャー,超獣の始動
+MQAS,虐食虫グリード・ワーム, パラサイトワーム, , C,3, 闇, , クリーチャー,超獣の始動
+PUAS,凶食虫スティンガーワーム, パラサイトワーム, , C,3, 闇, , クリーチャー,超獣の始動
+SYAS,ポイズン・ティアラ, ブレインジャッカー, , C,3, 闇, , クリーチャー,超獣の始動
+V4AS,捕らえる者ボーン・スパイダー, リビング・デッド, , C,3, 闇, , クリーチャー,超獣の始動
+AAAR,暗闇に潜む者バット・ドクター, ゴースト, , UC,3, 闇, , クリーチャー,超獣の始動
+DEAR,死胞虫エンリッチ・ワーム, パラサイトワーム, , UC,3, 闇, , クリーチャー,超獣の始動
+ZAAS,牙男, ヘドリアン, , C,4, 闇, , クリーチャー,超獣の始動
+4EAS,泥男, ヘドリアン, , C,4, 闇, , クリーチャー,超獣の始動
+GIAR,虚空の翼ダークモルダー, デーモン・コマンド, , UC,4, 闇, , クリーチャー,超獣の始動
+JMAR,カース・ペンダント, ブレインジャッカー, , UC,4, 闇, , クリーチャー,超獣の始動
+4EAP,呪いの影シャドウ・ムーン, ゴースト, , R,4, 闇, , クリーチャー,超獣の始動
+AAAQ,地獄の門番デスモーリー, デーモン・コマンド, , R,4, 闇, , クリーチャー,超獣の始動
+DEAQ,アンバー・ピアス, ブレインジャッカー, , R,4, 闇, , クリーチャー,超獣の始動
+4EAO,憤怒の猛将ダイダロス, デーモン・コマンド, , SR,4, 闇, , クリーチャー,超獣の始動
+ZAAR,死の宣告, 呪文, , UC,4, 闇, , 呪文,超獣の始動
+V4AQ,プライマス・スクリーム, 呪文, , R,4, 闇, , 呪文,超獣の始動
+MQAR,ギガベロス, キマイラ, , UC,5, 闇, , クリーチャー,超獣の始動
+PUAR,剣舞の修羅ヴァシュナ, デーモン・コマンド, , UC,5, 闇, , クリーチャー,超獣の始動
+GIAQ,砕骨の刺客ゾルバス, デーモン・コマンド, , R,5, 闇, , クリーチャー,超獣の始動
+SYAQ,悪魔神ロックデウス, デーモン・コマンド, , R,5, 闇, , クリーチャー,超獣の始動
+MQAP,屑男, ヘドリアン, , VR,5, 闇, , クリーチャー,超獣の始動
+V4AP,魔獣虫カオス・ワーム, パラサイトワーム, , VR,5, 闇, , クリーチャー,超獣の始動
+AAAT,ギガボアー, キマイラ, , C,6, 闇, , クリーチャー,超獣の始動
+SYAR,憎悪の騎士ガミル, デーモン・コマンド, , UC,6, 闇, , クリーチャー,超獣の始動
+JMAQ,邪妃グレゴリア, ダークロード, , R,6, 闇, , クリーチャー,超獣の始動
+MQAQ,冥将ダムド, ダークロード, , R,6, 闇, , クリーチャー,超獣の始動
+JMAP,腐食虫スワンプワーム, パラサイトワーム, , VR,6, 闇, , クリーチャー,超獣の始動
+ZAAP,デーモン・ハンド, 呪文, , VR,6, 闇, , 呪文,超獣の始動
+AAAP,魔刻の騎士オルゲイト, デーモン・コマンド, , SR,7, 闇, , クリーチャー,超獣の始動
+PUAQ,凶闘の魔神ギリエル, デーモン・コマンド, , R,8, 闇, , クリーチャー,超獣の始動
+ZAAQ,いけにえの鎖, 呪文, , R,8, 闇, , 呪文,超獣の始動
+PUAP,ギガルゴン, キマイラ, , VR,8, 闇, , クリーチャー,超獣の始動
+GIAP,ロスト・ソウル, 呪文, , SR,8, 闇, , 呪文,超獣の始動
+SYAP,妖姫シルフィ, ダークロード, , VR,8, 闇, , クリーチャー,超獣の始動
+DEAP,悪魔神バロム, デーモン・コマンド, , SR,8, 闇, , クリーチャー,超獣の始動
