+++ src/dmps_cards/series1/series1_blue_20200513.csv
@@ -1,0 +1,45 @@
+id,name,tribe1,tribe2,rarity,cost,culture1,culture2,category,series
+PUAN,ゼピメテウス, シー・ハッカー, , C,1, 水, , クリーチャー,超獣の始動
+SYAN,エレガント・ランプ, サイバー・ウイルス, , C,2, 水, , クリーチャー,超獣の始動
+V4AN,狩猟魚, フィッシュ, , C,2, 水, , クリーチャー,超獣の始動
+ZAAN,アルオム, サイバーロード, , C,2, 水, , クリーチャー,超獣の始動
+ZAAO,エナジー・ライト, 呪文, , C,3, 水, , 呪文,超獣の始動
+JMAN,ジーニアス・ディザスター, 呪文, , UC,3, 水, , 呪文,超獣の始動
+DEAM,ミラクル・サーチャー, 呪文, , R,3, 水, , 呪文,超獣の始動
+4EAN,キング・クラーケン, リヴァイアサン, , C,3, 水, , クリーチャー,超獣の始動
+AAAO,アクア・ソルジャー, リキッド・ピープル, , C,3, 水, , クリーチャー,超獣の始動
+DEAO,アクア・チャージャー, リキッド・ピープル, , C,3, 水, , クリーチャー,超獣の始動
+PUAM,アングラー・クラスター, サイバー・クラスター, , UC,3, 水, , クリーチャー,超獣の始動
+MQAL,ストリーミング・ビジョン, 呪文, , VR,4, 水, , 呪文,超獣の始動
+MQAN,ディープ・オペレーション, 呪文, , UC,4, 水, , 呪文,超獣の始動
+V4AK,クリスタル・パラディン, リキッド・ピープル, , SR,4, 水, , クリーチャー,超獣の始動
+4EAK,マーチング・スプライト, サイバー・ウイルス, , SR,4, 水, , クリーチャー,超獣の始動
+AAAL,ラプター・フィッシュ, ゲル・フィッシュ, , VR,4, 水, , クリーチャー,超獣の始動
+GIAO,シザー・アイ, ゲル・フィッシュ, , C,4, 水, , クリーチャー,超獣の始動
+JMAO,リボルバー・フィッシュ, ゲル・フィッシュ, , C,4, 水, , クリーチャー,超獣の始動
+MQAO,フェアリー・チャイルド, サイバー・ウイルス, , C,4, 水, , クリーチャー,超獣の始動
+PUAO,ペコタン, サイバーロード, , C,4, 水, , クリーチャー,超獣の始動
+SYAM,ストーム・クロウラー, アースイーター, , UC,4, 水, , クリーチャー,超獣の始動
+PUAL,一角魚, フィッシュ, , R,4, 水, , クリーチャー,超獣の始動
+SYAO,トロピコ, サイバーロード, , C,5, 水, , クリーチャー,超獣の始動
+V4AM,キング・ヘルゴラント, リヴァイアサン, , UC,5, 水, , クリーチャー,超獣の始動
+ZAAM,アクア・グラップラー, リキッド・ピープル, , UC,5, 水, , クリーチャー,超獣の始動
+DEAN,キング・ケール, リヴァイアサン, , UC,5, 水, , クリーチャー,超獣の始動
+SYAL,ソーサーヘッド・シャーク, ゲル・フィッシュ, , R,5, 水, , クリーチャー,超獣の始動
+V4AL,ミラージュ・マーメイド, ゲル・フィッシュ, , R,5, 水, , クリーチャー,超獣の始動
+GIAM,テレポーテーション, 呪文, , R,5, 水, , 呪文,超獣の始動
+JMAM,トリプル・ブレイン, 呪文, , R,5, 水, , 呪文,超獣の始動
+PUAK,アクア・サーファー, リキッド・ピープル, , VR,6, 水, , クリーチャー,超獣の始動
+DEAL,プラズマ・チェイサー, ゲル・フィッシュ, , VR,6, 水, , クリーチャー,超獣の始動
+JMAL,クリスタル・ランサー, リキッド・ピープル, , VR,6, 水, , クリーチャー,超獣の始動
+V4AO,シーマイン, フィッシュ, , C,6, 水, , クリーチャー,超獣の始動
+4EAM,キング・ネプタス, リヴァイアサン, , UC,6, 水, , クリーチャー,超獣の始動
+AAAN,アクア・レンジャー, リキッド・ピープル, , UC,6, 水, , クリーチャー,超獣の始動
+ZAAL,キング・ポセイドン, リヴァイアサン, , R,6, 水, , クリーチャー,超獣の始動
+GIAL,キング・マゼラン, リヴァイアサン, , VR,7, 水, , クリーチャー,超獣の始動
+ZAAK,レジェンダリー・パイロン, リヴァイアサン, , SR,7, 水, , クリーチャー,超獣の始動
+MQAM,マリン・スクランブル, 呪文, , R,7, 水, , 呪文,超獣の始動
+4EAL,混沌魚, ゲル・フィッシュ, , R,7, 水, , クリーチャー,超獣の始動
+AAAM,キング・ノーチラス, リヴァイアサン, , R,7, 水, , クリーチャー,超獣の始動
+SYAK,アクア・スナイパー, リキッド・ピープル, , SR,8, 水, , クリーチャー,超獣の始動
+GIAN,シー・スラッグ, ゲル・フィッシュ, , UC,8, 水, , クリーチャー,超獣の始動
