+++ src/dmps_cards/series1/series1_green_20200513.csv
@@ -1,0 +1,45 @@
+id,name,tribe1,tribe2,rarity,cost,culture1,culture2,category,series
+V4AZ,スナイプ・モスキート, ギガント・インセクト, , UC,1, 自然, , クリーチャー,超獣の始動
+MQA2,大地の咆哮, 呪文, , UC,2, 自然, , 呪文,超獣の始動
+V4A2,アンビション・ビー, ギガント・インセクト, , C,2, 自然, , クリーチャー,超獣の始動
+ZAAZ,シビレアシダケ, バルーン・マッシュルーム, , UC,2, 自然, , クリーチャー,超獣の始動
+ZAA2,原子の棍棒, ビーストフォーク, , C,2, 自然, , クリーチャー,超獣の始動
+4EA2,孤高の願, ビーストフォーク, , C,2, 自然, , クリーチャー,超獣の始動
+PUAZ,大勇者「大地の猛攻」, ビーストフォーク, , R,2, 自然, , クリーチャー,超獣の始動
+PUA2,ディンジョン・ゲート, 呪文, , UC,3, 自然, , 呪文,超獣の始動
+AAA3,ブルーレイザー・ビートル, ギガント・インセクト, , C,3, 自然, , クリーチャー,超獣の始動
+DEA3,ランチャー・スパイダー, ギガント・インセクト, , C,3, 自然, , クリーチャー,超獣の始動
+GIA3,白銀の牙, ビーストフォーク, , C,3, 自然, , クリーチャー,超獣の始動
+JMA3,無敵の咆哮, ビーストフォーク, , C,3, 自然, , クリーチャー,超獣の始動
+DEAY,無垢の宝剣, ビーストフォーク, , VR,4, 自然, , クリーチャー,超獣の始動
+AAA4,オーラ・ブラスター, 呪文, , C,4, 自然, , 呪文,超獣の始動
+SYA2,再誕の森, 呪文, , UC,4, 自然, , 呪文,超獣の始動
+4EAZ,マボロシキリダケ, バルーン・マッシュルーム, , UC,4, 自然, , クリーチャー,超獣の始動
+V4AY,秋風妖精リップル, スノーフェアリー, , R,4, 自然, , クリーチャー,超獣の始動
+ZAAY,念仏エルフィン, ツリーフォーク, , R,4, 自然, , クリーチャー,超獣の始動
+MQA3,シェル・キャノン, コロニー・ビートル, , C,4, 自然, , クリーチャー,超獣の始動
+PUA3,ハザード・ホッパー, ギガント・インセクト, , C,4, 自然, , クリーチャー,超獣の始動
+SYA3,銀の拳, ビーストフォーク, , C,4, 自然, , クリーチャー,超獣の始動
+V4A3,激闘するフェザー・ホーン, ホーン・ビースト, , C,4, 自然, , クリーチャー,超獣の始動
+ZAA3,からみカズラ, ツリーフォーク, , C,4, 自然, , クリーチャー,超獣の始動
+4EAX,大昆虫ジェネラルマンティス, ギガント・インセクト, , SR,5, 自然, , クリーチャー,超獣の始動
+GIAY,天空の超人, ジャイアント, , VR,5, 自然, , クリーチャー,超獣の始動
+JMAY,鳴動するギガ・ホーン, ホーン・ビースト, , VR,5, 自然, , クリーチャー,超獣の始動
+MQAY,鼓舞の化身, ミステリー・トーテム, , VR,5, 自然, , クリーチャー,超獣の始動
+SYAZ,アルティメット・フォース, 呪文, , R,5, 自然, , 呪文,超獣の始動
+AAA2,アンブッシュ・スコーピオン, ギガント・インセクト, , UC,5, 自然, , クリーチャー,超獣の始動
+DEA2,奇面ざくろ, ツリーフォーク, , UC,5, 自然, , クリーチャー,超獣の始動
+4EAY,ブラッドウイング・マンティス, ギガント・インセクト, , R,5, 自然, , クリーチャー,超獣の始動
+AAAZ,激昂するダッシュ・ホーン, ホーン・ビースト, , R,5, 自然, , クリーチャー,超獣の始動
+AAAY,大勇者「ふたつ牙」, ビーストフォーク, , SR,6, 自然, , クリーチャー,超獣の始動
+PUAY,オブシディアン・ビートル, ギガント・インセクト, , VR,6, 自然, , クリーチャー,超獣の始動
+SYAY,ナチュラル・トラップ, 呪文, , VR,6, 自然, , 呪文,超獣の始動
+DEAZ,神拳の超人, ジャイアント, , R,6, 自然, , クリーチャー,超獣の始動
+4EA3,シェル・タワー, コロニー・ビートル, , C,6, 自然, , クリーチャー,超獣の始動
+V4AX,二角の超人, ジャイアント, , SR,7, 自然, , クリーチャー,超獣の始動
+GIA2,シェル・ストーム, コロニー・ビートル, , UC,7, 自然, , クリーチャー,超獣の始動
+ZAAX,不落の超人, ジャイアント, , SR,8, 自然, , クリーチャー,超獣の始動
+JMA2,シザーズ・ビートル, ギガント・インセクト, , UC,7, 自然, , クリーチャー,超獣の始動
+GIAZ,夜明けの超人, ジャイアント, , R,7, 自然, , クリーチャー,超獣の始動
+JMAZ,ゼノ・マンティス, ギガント・インセクト, , R,7, 自然, , クリーチャー,超獣の始動
+MQAZ,ガイアクロウ・ワスプ, ギガント・インセクト, , R,7, 自然, , クリーチャー,超獣の始動
