+++ src/dmps_cards/series1/series1_red_20200513.csv
@@ -1,0 +1,45 @@
+id,name,tribe1,tribe2,rarity,cost ,culture1,culture2,category,series
+GIAW,火炎流星弾, 呪文, , UC,1, 火, , 呪文,超獣の始動
+JMAW,凶戦士ブレイズ・クロー, ドラゴノイド, , C,1, 火, , クリーチャー,超獣の始動
+MQAW,小さな勇者ゲット, ヒューマノイド, , C,2, 火, , クリーチャー,超獣の始動
+JMAV,奇襲兵ブルレイザー, ドラゴノイド, , UC,2, 火, , クリーチャー,超獣の始動
+MQAV,無謀鉄人チョイヤ, ヒューマノイド, , UC,2, 火, , クリーチャー,超獣の始動
+SYAV,リッピ・ポルタ, ファイアー・バード, , UC,2, 火, , クリーチャー,超獣の始動
+PUAX,クリムゾン・ハンマー, 呪文, , C,2, 火, , 呪文,超獣の始動
+PUAW,突神兵ドガーン, アーマロイド, , C,3, 火, , クリーチャー,超獣の始動
+SYAW,騎兵総長キュラトプス, ドラゴノイド, , C,3, 火, , クリーチャー,超獣の始動
+V4AW,一撃必殺のホーバス, ヒューマノイド, , C,3, 火, , クリーチャー,超獣の始動
+ZAAW,チッチ・ホッピー, ファイアー・バード, , C,3, 火, , クリーチャー,超獣の始動
+PUAV,コッコ・ルピア, ファイアー・バード, , UC,3, 火, , クリーチャー,超獣の始動
+V4AV,解体屋ピーカプ, マシン・イーター, , UC,3, 火, , クリーチャー,超獣の始動
+ZAAV,クレーターザウルス, ロック・ビースト, , UC,3, 火, , クリーチャー,超獣の始動
+4EAU,ボルカニック・アロー, 呪文, , R,2, 火, , 呪文,超獣の始動
+JMAU,トット・ピピッチ, ファイアー・バード, , R,3, 火, , クリーチャー,超獣の始動
+4EAW,撃神兵マシンガー, アーマロイド, , C,4, 火, , クリーチャー,超獣の始動
+AAAX,砲神兵バゴーン, アーマロイド, , C,4, 火, , クリーチャー,超獣の始動
+DEAX,超砲手ボルカノドン, ドラゴノイド, , C,4, 火, , クリーチャー,超獣の始動
+GIAX,カルドザウルス, ロック・ビースト, , C,4, 火, , クリーチャー,超獣の始動
+AAAV,マグマ・ゲイザー, 呪文, , R,3, 火, , 呪文,超獣の始動
+MQAU,メテオザウルス, ロック・ビースト, , R,4, 火, , クリーチャー,超獣の始動
+JMAX,猛爆軍曹ボンバット, ドラゴノイド, , C,5, 火, , クリーチャー,超獣の始動
+SYAX,クエイク・ゲート, 呪文, , C,4, 火, , 呪文,超獣の始動
+DEAV,灼熱波, 呪文, , R,5, 火, , 呪文,超獣の始動
+MQAX,ガルザウルス, ロック・ビースト, , C,5, 火, , クリーチャー,超獣の始動
+4EAV,ガザリアス・ドラゴン, アーマード・ドラゴン, , UC,5, 火, , クリーチャー,超獣の始動
+AAAW,放浪の勇者ジージョ, マシン・イーター, , UC,5, 火, , クリーチャー,超獣の始動
+DEAW,マグマティラノス, ロック・ビースト, , UC,5, 火, , クリーチャー,超獣の始動
+ZAAU,機神装甲ヴァルカイザー, ヒューマノイド, , R,5, 火, , クリーチャー,超獣の始動
+PUAU,怪神兵ナグール, アーマロイド, , R,6, 火, , クリーチャー,超獣の始動
+SYAU,勇神兵エグゾリウス, アーマロイド, , R,6, 火, , クリーチャー,超獣の始動
+V4AT,襲撃者エグゼドライブ, ドラゴノイド, , VR,3, 火, , クリーチャー,超獣の始動
+SYAT,機神装甲ヴァルボーグ, ヒューマノイド, , SR,3, 火, , クリーチャー,超獣の始動
+V4AU,メテオ・ドラゴン, アーマード・ドラゴン, , R,7, 火, , クリーチャー,超獣の始動
+GIAV,ドリル・トラップ, 呪文, , R,7, 火, , 呪文,超獣の始動
+DEAU,超巨岩獣ドボルカイザー, ロック・ビースト, , VR,6, 火, , クリーチャー,超獣の始動
+GIAU,バースト・ショット, 呪文, , VR,6, 火, , 呪文,超獣の始動
+JMAT,ボルメテウス・ホワイト・ドラゴン, アーマード・ドラゴン, , SR,7, 火, , クリーチャー,超獣の始動
+MQAT,ツインキャノン・ワイバーン, アーマード・ワイバーン, , SR,7, 火, , クリーチャー,超獣の始動
+ZAAT,バルキリー・ドラゴン, アーマード・ドラゴン, , VR,7, 火, , クリーチャー,超獣の始動
+4EAT,ミラフォース・ドラゴン, アーマード・ドラゴン, , VR,7, 火, , クリーチャー,超獣の始動
+AAAU,ガトリング・ワイバーン, アーマード・ワイバーン, , VR,7, 火, , クリーチャー,超獣の始動
+PUAT,クリムゾン・ワイバーン, アーマード・ワイバーン, , SR,8, 火, , クリーチャー,超獣の始動
