+++ src/dmps_cards/series1/series1_white_20200513.csv
@@ -1,0 +1,45 @@
+id,name,tribe1,tribe2,rarity,cost,culture1,culture2,category,series
+MQAK,プロテクト・フォース, 呪文, , C,1, 光, , 呪文,超獣の始動
+DEAJ,予言者ジェス, ライトブリンガー, , C,1, 光, , クリーチャー,超獣の始動
+ZAAI,フル・ディフェンサー, 呪文, , UC,2, 光, , 呪文,超獣の始動
+PUAJ,予言者リュゾル, ライトブリンガー, , C,2, 光, , クリーチャー,超獣の始動
+MQAJ,予言者カティノ, ライトブリンガー, , C,2, 光, , クリーチャー,超獣の始動
+JMAJ,碧玉草, スターライト・ツリー, , C,2, 光, , クリーチャー,超獣の始動
+GIAJ,牽制の使徒カルエス, イニシエート, , C,2, 光, , クリーチャー,超獣の始動
+ZAAJ,乱雲の使徒カルゴ, イニシエート, , C,3, 光, , クリーチャー,超獣の始動
+V4AJ,雷光の使徒ミール, イニシエート, , C,3, 光, , クリーチャー,超獣の始動
+SYAJ,弾丸の使徒イーレ, イニシエート, , C,3, 光, , クリーチャー,超獣の始動
+DEAI,紅玉草, スターライト・ツリー, , UC,3, 光, , クリーチャー,超獣の始動
+AAAK,予言者べクス, ライトブリンガー, , C,3, 光, , クリーチャー,超獣の始動
+AAAJ,ロジック・キューブ, 呪文, , UC,3, 光, , 呪文,超獣の始動
+4EAJ,翡翠樹, スターライト・ツリー, , C,3, 光, , クリーチャー,超獣の始動
+4EAI,エンジェル・ソング, 呪文, , UC,3, 光, , 呪文,超獣の始動
+4EAG,守護聖天ラルバ・ギア, ガーディアン, , VR,3, 光, , クリーチャー,超獣の始動
+ZAAH,グローリー・スノー, 呪文, , R,4, 光, , 呪文,超獣の始動
+JMAK,陽光の求道者ル・パーレ, メガサンダー, , C,4, 光, , クリーチャー,超獣の始動
+JMAI,光魂の伝道士クルス, バーサーカー, , UC,4, 光, , クリーチャー,超獣の始動
+GIAK,月光の守護者ディア・ノーク, ガーディアン, , C,4, 光, , クリーチャー,超獣の始動
+GIAI,閃光の使徒レーナ, イニシエート, , UC,4, 光, , クリーチャー,超獣の始動
+DEAK,捕縛の使徒ローム, イニシエート, , C,4, 光, , クリーチャー,超獣の始動
+DEAH,光臨の精霊ピカリエ, エンジェル・コマンド, , R,4, 光, , クリーチャー,超獣の始動
+4EAH,ホーリー・メール, 呪文, , R,4, 光, , 呪文,超獣の始動
+V4AH,聖霊王アヴァラス, エンジェル・コマンド, , R,5, 光, , クリーチャー,超獣の始動
+PUAG,真理の伝道師レイーラ, バーサーカー, , VR,5, 光, , クリーチャー,超獣の始動
+MQAI,希望の使徒トール, イニシエート, , UC,5, 光, , クリーチャー,超獣の始動
+MQAH,閃光の伝道師ラグナ, バーサーカー, , R,5, 光, , クリーチャー,超獣の始動
+JMAH,暁の守護者ファルイーガ, ガーディアン, , R,5, 光, , クリーチャー,超獣の始動
+GIAH,青嵐の精霊バルキア, エンジェル・コマンド, , R,5, 光, , クリーチャー,超獣の始動
+DEAG,雷鳴の守護者ミスト・リエス, ガーディアン, , SR,5, 光, , クリーチャー,超獣の始動
+AAAI,ダイヤモンド・カッター, 呪文, , R,5, 光, , 呪文,超獣の始動
+V4AI,雷電の求道者ラ・ビュー, メガサンダー, , UC,6, 光, , クリーチャー,超獣の始動
+SYAI,雷撃の求道者ラ・ガイル, メガサンダー, , UC,6, 光, , クリーチャー,超獣の始動
+SYAG,栄光の精霊アイリス, エンジェル・コマンド, , VR,6, 光, , クリーチャー,超獣の始動
+PUAI,天空の守護者グラン・ギューレ, ガーディアン, , UC,6, 光, , クリーチャー,超獣の始動
+MQAG,精霊王アルカディアス, エンジェル・コマンド, , SR,6, 光, , クリーチャー,超獣の始動
+AAAH,ホーリー・スパーク, 呪文, , VR,6, 光, , 呪文,超獣の始動
+V4AG,風撃の求道者ラ・バイル, メガサンダー, , VR,7, 光, , クリーチャー,超獣の始動
+PUAH,聖弓の求道者レ・ビール, メガサンダー, , R,7, 光, , クリーチャー,超獣の始動
+GIAG,光翼の精霊サイフォス, エンジェル・コマンド, , SR,7, 光, , クリーチャー,超獣の始動
+ZAAG,閃光の求道者ラ・ベイル, メガサンダー, , VR,8, 光, , クリーチャー,超獣の始動
+SYAH,流星の精霊ミーア, エンジェル・コマンド, , R,8, 光, , クリーチャー,超獣の始動
+JMAG,聖核の精霊ウルセリオス, エンジェル・コマンド, , SR,8, 光, , クリーチャー,超獣の始動
